Add GUI
This is a massive commit that adds the following functionality
* 2 new GUI windows
** the main window displays the list of actions to do
** the add_action window is (currently) hard coded to add a mouse click action
* ties the GUI together with the mouse listener to capture mouse press events
* allows the ability to add mouse click actions

In addition to all of the above functionality, it's important to note that a MVC design has been created.  The view utilizes the QT framework (PySide2) and the model currently is not persisted after each execution of the program (i.e. there is no back end yet).